requirements.txt
flask
openai
autogen-agentchat[retrievechat]~=0.2